+++ OggConverter/src/Misc/Utilities.cs
@@ -43,14 +43,15 @@ namespace OggConverter
         /// <summary>
         /// Returns MSC Music Manager version
         /// </summary>
+        /// <param name="fullVersion">If true, returns the name with inclusion of revision number.</param>
         /// <returns></returns>
         public static string GetVersion(bool fullVersion = false)
         {
             Version version = Assembly.GetExecutingAssembly().GetName().Version;
+
             if (fullVersion)
-            {
                 return $"{version.Major}.{version.Minor}.{version.Build}.{version.Revision}";
-            }
+
             return version.Build == 0 ? $"{version.Major}.{version.Minor}" : $"{version.Major}.{version.Minor}.{version.Build}";
         }
 
@@ -60,17 +61,17 @@ namespace OggConverter
                 $"MSC Music Manager uses FFmpeg, which is licensed under LGPL 2.1 license.";
 
         /// <summary>
-        /// Checks if file is being used by other process and returns the value.
+        /// Checks if file is being used by some other process.
         /// </summary>
-        /// <param name="filename">Path to the file</param>
-        /// <returns>If any process is in FileStream, returns true (as it's used). Else it returns false (file is free to use)</returns>
-        public static bool IsFileReady(string filename)
+        /// <param name="fileName">Path to the file</param>
+        /// <returns>If any process is in FileStream, returns true (as it's used). Else it returns false (file is unlocked)</returns>
+        public static bool IsFileReady(string fileName)
         {
             // If the file can be opened for exclusive access it means that the file
             // is no longer locked by another process.
             try
             {
-                using (FileStream inputStream = File.Open(filename, FileMode.Open, FileAccess.Read, FileShare.None))
+                using (FileStream inputStream = File.Open(fileName, FileMode.Open, FileAccess.Read, FileShare.None))
                     return inputStream.Length > 0;
             }
             catch
@@ -79,6 +80,9 @@ namespace OggConverter
             }
         }
 
+        /// <summary>
+        /// List of files to delete (old unused files, temps, etc.)
+        /// </summary>
         static readonly string[] filesToDelete = new string[]
         {
             "NReco.VideoConverter.dll", "updater.bat", "restart.bat",